+++ addon/CHANGELOG.md
@@ -1,4 +1,4 @@
-##### *GENERATED FILE [2026-07-21 21:11 UTC] — do not edit by hand.*
+##### *GENERATED FILE [2026-07-22 20:55 UTC] — do not edit by hand.*
 
 # Changelog